Fix doubles().fromZeroToOne() to generate uniformly distributed values
`SourceDSL.doubles().fromZeroToOne()` was generating mostly really
small numbers. It was because generation was touching exponents part
of double precision representation (bits 52 to 63), but in a way where
it only generated zero or negative exponents. This code fixes it as now
we only generate fraction part of double (bits 0 to 51).